+++ src/Checkers/ChangedChecker/ChangeWatcher.php
@@ -30,6 +30,10 @@ class ChangeWatcher
         $command = sprintf('git log --oneline -n %d', $commits);
         $commits = shell_exec($command);
         $lastResult = $this->getLastCommit();
+        if (empty($lastResult)) {
+            $this->setChangelogChanged();
+            return $this->changelogChanged();
+        }
         if (!strpos($commits, $lastResult)) {
             $this->setChangelogChanged();
             return $this->changelogChanged();