+++ src/Checkers/TypeChecker.php
@@ -2,26 +2,25 @@
 
 namespace Boscho87\ChangelogChecker\Checkers;
 
+use Boscho87\ChangelogChecker\Changelog\Regex;
+
 /**
  * Class TypeChecker
  */
 class TypeChecker extends AbstractChecker
 {
-    private array $allowedTypes = ['Added', 'Changed', 'Deprecated', 'Removed', 'Fixed', 'Security'];
-
-
     protected function check(): void
     {
         foreach ($this->file as $line) {
             if ($this->isTypeLine($line)) {
-                preg_match($this->getAllowedTypePattern(), $line, $matches);
+                preg_match(Regex::allowedTypesPattern(), $line, $matches);
                 $typeString = $matches[2] ?? '';
-                if (!in_array($typeString, $this->allowedTypes)) {
+                if (!in_array($typeString, Regex::$allowedTypes)) {
                     $this->addErrorMessage(sprintf(
                         'Line %s has invalid Type: "%s", allowed are %s',
                         $this->file->lineNumber(),
                         $line,
-                        implode(',', $this->allowedTypes)
+                        implode(',', Regex::$allowedTypes)
                     ));
                 }
             }
@@ -30,13 +29,6 @@ class TypeChecker extends AbstractChecker
 
     protected function fix(): void
     {
-        // TODO: Implement fix() method.
-    }
-
-
-    private function getAllowedTypePattern(): string
-    {
-        $types = implode('|', $this->allowedTypes);
-        return sprintf('/(###\s)(%s)$/', $types);
+        //no fix possible for this case
     }
 }