+++ src/Composer/Regex.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Boscho87\ChangelogChecker\Composer;
+
+/**
+ * Class Regex
+ */
+final class Regex
+{
+    public const COMPOSER_VERSION = '/"version":"\d+\.\d+\.\d+"/';
+}