+++ src/Exception/InvalidChangeTypeException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Boscho87\ChangelogChecker\Exception;
+
+/**
+ * Class InvalidChangeTypeException
+ */
+class InvalidChangeTypeException extends ChangelogCheckerException
+{
+}