+++ src/Exception/InvalidInstanceException.php
@@ -2,6 +2,9 @@
 
 namespace Boscho87\ChangelogChecker\Exception;
 
+/**
+ * Class InvalidInstanceException
+ */
 class InvalidInstanceException extends ChangelogCheckerException
 {
 }